@@ -7,7 +7,7 @@
 
     <groupId>io.github.kinyha</groupId>  <!-- Твой GitHub username -->
     <artifactId>http-request-logger-parent</artifactId>
-    <version>1.0.0</version>
+    <version>1.0.2</version>  <!-- Новая версия для Spring Boot 3 -->
     <packaging>pom</packaging>
 
     <!-- Новое - добавь это -->
@@ -24,10 +24,8 @@
     <url>https://github.com/kinyha/http-request-logger</url>
 
     <properties>
-        <java.version>11</java.version>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
-        <spring-boot.version>2.7.14</spring-boot.version>
+        <spring-boot.version>3.2.0</spring-boot.version>
+        <java.version>17</java.version>
     </properties>
 
     <modules>
