@@ -1,5 +1,7 @@
 package io.github.kinyha.requestlogger.autoconfigure;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -9,7 +11,7 @@ import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import javax.servlet.Filter;
+import jakarta.servlet.Filter;
 
 @Configuration
 @ConditionalOnWebApplication
@@ -23,10 +25,15 @@ import javax.servlet.Filter;
 @EnableConfigurationProperties(HttpRequestLoggerProperties.class)
 public class HttpRequestLoggerAutoConfiguration {
 
+    private static final Logger log = LoggerFactory.getLogger(HttpRequestLoggerAutoConfiguration.class);
+
     @Bean
     @ConditionalOnMissingBean
     public FilterRegistrationBean<HttpRequestLoggingFilter> httpRequestLoggingFilter(
             HttpRequestLoggerProperties properties) {
+
+        log.info("=== CREATING HTTP REQUEST LOGGING FILTER ==="); // Добавь это
+        log.info("Properties: enabled={}, includePayload={}", properties.isEnabled(), properties.isIncludePayload());
 
         FilterRegistrationBean<HttpRequestLoggingFilter> registrationBean = new FilterRegistrationBean<>();
         registrationBean.setFilter(new HttpRequestLoggingFilter(properties));
